Refactor bus stop and bus stop group APIs
- Changed JSON structure in bus_stop_groups.json from "stops" to "busStops".
- Updated bus_stops.json to correct the naming of a bus stop.
- Modified middleware to handle service errors more appropriately.
- Renamed server methods for bus stop groups to follow a consistent naming convention.
- Updated domain models to use float64 for latitude and longitude.
- Adjusted handler logic to accommodate changes in data types.
- Updated OpenAPI models to reflect changes in parameter names and types.
- Created new TypeSpec files for bus stop groups and bus stop timetables.
- Removed redundant code from busStops.tsp and organized routes into separate files.
- Updated generated OpenAPI definitions to match the new API structure and naming conventions.

+++ apps/api/internal/handler/bus_stop_handler.go
@@ -113,11 +113,11 @@ func (h *BusStopHandler) GetBusStopTimetable(ctx echo.Context, id int32, date oa
 		Segments: []oapi.ModelsBusStopSegment{},
 	}
 	if busStop.Lat != nil {
-		lat := float32(*busStop.Lat)
+		lat := float64(*busStop.Lat)
 		timetable.Lat = lat
 	}
 	if busStop.Lng != nil {
-		lon := float32(*busStop.Lng)
+		lon := float64(*busStop.Lng)
 		timetable.Lon = lon
 	}
